+++ docs/js/picker.en-GB.js
@@ -1,11 +1,10 @@
 (function (global, factory) {
-  typeof exports === 'object' && typeof module !== 'undefined' ? factory(require('pickerjs')) :
-  typeof define === 'function' && define.amd ? define(['pickerjs'], factory) :
-  (factory(global.Picker));
-}(this, (function (Picker) {
+  typeof exports === 'object' && typeof module !== 'undefined' ? module.exports = factory() :
+  typeof define === 'function' && define.amd ? define(factory) : factory(global.Picker);
+}(typeof self !== 'undefined' ? self : this, function (Picker) {
   'use strict';
 
   Picker.languages['en-GB'] = {
     format: 'DD/MM/YYYY HH:mm'
   };
-})));
+}));